Lots of small changes, mostly having to do with porting to 64 bits. - Added check for ptrdiff_t (type defined to hold difference of   pointers) with fallback to int. - Use void * to hold stream_data in stream.mx. - Added symbol SIZEOF_OID for when oid is unsigned long instead of   unsigned int. - Added some alternative code for when SIZEOF_OID != SIZEOF_INT (not   really tested). - Removed some unnecessary casts (casting values to a type they   already have). - Changed various int variables to oid since they are used to hold   results of calculations with oid values (important when sizeof(oid)   > sizeof(int)). - Removed declarations of some unused variables. - Fixed arguments and format strings for *printf and *scanf: In C99   (the latest ISO C standard) you can have %td, %zd, and %p for args   of type ptrdiff_t, size_t, and pointer respectively--use that, but   give alternatives; use long or unsgigned long in other cases.   Also removed \ in front of @ in format strings. - Changed type MT_Id to unsigned long from void * (matches better with   the way this type is used). - Removed LDFLAGS = -all-static in mel Makefile.ag.  It doesn't seem   to be necessary, and it causes complaints from the native Windows   compiler. - Removed some degenerate tests: comparing an unsigned value for < 0   or >= 0 gives a predictable result (plus a warning from some   compilers).  Note, oid is an unsigned value. - Changed comparison in enum.mx from xx>MAX_ENUM to xx>=MAX_ENUM,   since xx is used to index an array of size MAX_ENUM. - Changed use of mallinfo struct values in mprof.mx since these values   are not guaranteed to be int (they are long on 64-bit IRIX).

+++ MonetDB/src/testing/difflib.c
@@ -198,8 +198,8 @@ int lw_diff2wc_diff (int doChar, char *lw_diff_fn, char *wc_diff_fn)
   if(!(ok=fgets(line,BUFLEN,lw_diff_fp)))
     { fclose(lw_diff_fp); ERRHNDL(0,"empty file in lw_diff2wc_diff:",lw_diff_fn,1); }
   
-  sprintf(wc_old_fn,"%s%c.difflib-%ld-lw_diff2wc_diff-old",tmpdir(),DIR_SEP,getpid()); fn[0]=wc_old_fn;
-  sprintf(wc_new_fn,"%s%c.difflib-%ld-lw_diff2wc_diff-new",tmpdir(),DIR_SEP,getpid()); fn[1]=wc_new_fn;
+  sprintf(wc_old_fn,"%s%c.difflib-%ld-lw_diff2wc_diff-old",tmpdir(),DIR_SEP,(long)getpid()); fn[0]=wc_old_fn;
+  sprintf(wc_new_fn,"%s%c.difflib-%ld-lw_diff2wc_diff-new",tmpdir(),DIR_SEP,(long)getpid()); fn[1]=wc_new_fn;
   
   wc_diff_fp=Wfopen(wc_diff_fn);
   while(ok&&strncmp(line,"@@ -",4)) { fprintf(wc_diff_fp,"%s",line); ok=fgets(line,BUFLEN,lw_diff_fp); }
@@ -330,7 +330,7 @@ int l_diff2c_diff (char *l_diff_fn, char *c_diff_fn)
 
   TRACE(fprintf(STDERR,"l_diff2c_diff(%s,%s)\n",l_diff_fn,c_diff_fn));
 
-  sprintf(w_diff_fn,"%s%c.difflib-%ld-l_diff2c_diff-w_diff",tmpdir(),DIR_SEP,getpid());
+  sprintf(w_diff_fn,"%s%c.difflib-%ld-l_diff2c_diff-w_diff",tmpdir(),DIR_SEP,(long)getpid());
 
   if(!l_diff2w_diff(l_diff_fn,w_diff_fn))
     { UNLINK(w_diff_fn); ERRHNDL(0,"l_diff2w_diff returns 0 in l_diff2c_diff","",1); }
@@ -349,7 +349,7 @@ int u_diff2w_diff (char *u_diff_fn, char *w_diff_fn)
 
   TRACE(fprintf(STDERR,"u_diff2w_diff(%s,%s)\n",u_diff_fn,w_diff_fn));
 
-  sprintf(l_diff_fn,"%s%c.difflib-%ld-u_diff2w_diff-l_diff",tmpdir(),DIR_SEP,getpid());
+  sprintf(l_diff_fn,"%s%c.difflib-%ld-u_diff2w_diff-l_diff",tmpdir(),DIR_SEP,(long)getpid());
 
   if(!u_diff2l_diff(u_diff_fn,l_diff_fn))
     { UNLINK(l_diff_fn); ERRHNDL(0,"u_diff2l_diff returns 0 in u_diff2w_diff","",1); }
@@ -368,7 +368,7 @@ int u_diff2c_diff (char *u_diff_fn, char *c_diff_fn)
 
   TRACE(fprintf(STDERR,"u_diff2c_diff(%s,%s)\n",u_diff_fn,c_diff_fn));
 
-  sprintf(w_diff_fn,"%s%c.difflib-%ld-u_diff2c_diff-w_diff",tmpdir(),DIR_SEP,getpid());
+  sprintf(w_diff_fn,"%s%c.difflib-%ld-u_diff2c_diff-w_diff",tmpdir(),DIR_SEP,(long)getpid());
 
   if(!u_diff2w_diff(u_diff_fn,w_diff_fn))
     { UNLINK(w_diff_fn); ERRHNDL(0,"u_diff2w_diff returns 0 in u_diff2c_diff","",1); }
@@ -403,7 +403,7 @@ int oldnew2l_diff (int context, char *ignore, char *old_fn, char *new_fn, char *
 
   TRACE(fprintf(STDERR,"oldnew2l_diff(%i,%s,%s,%s,%s)\n",context,ignore,old_fn,new_fn,l_diff_fn));
 
-  sprintf(u_diff_fn,"%s%c.difflib-%ld-oldnew2l_diff-u_diff",tmpdir(),DIR_SEP,getpid());
+  sprintf(u_diff_fn,"%s%c.difflib-%ld-oldnew2l_diff-u_diff",tmpdir(),DIR_SEP,(long)getpid());
 
   if(!oldnew2u_diff(context,ignore,old_fn,new_fn,u_diff_fn))
     { UNLINK(u_diff_fn); ERRHNDL(0,"oldnew2u_diff returns 0 in oldnew2l_diff","",1); }
@@ -422,7 +422,7 @@ int oldnew2w_diff (int context, char *ignore, char *old_fn, char *new_fn, char *
 
   TRACE(fprintf(STDERR,"oldnew2w_diff(%i,%s,%s,%s,%s)\n",context,ignore,old_fn,new_fn,w_diff_fn));
 
-  sprintf(l_diff_fn,"%s%c.difflib-%ld-oldnew2w_diff-l_diff",tmpdir(),DIR_SEP,getpid());
+  sprintf(l_diff_fn,"%s%c.difflib-%ld-oldnew2w_diff-l_diff",tmpdir(),DIR_SEP,(long)getpid());
 
   if(!oldnew2l_diff(context,ignore,old_fn,new_fn,l_diff_fn))
     { UNLINK(l_diff_fn); ERRHNDL(0,"oldnew2l_diff returns 0 in oldnew2w_diff","",1); }
@@ -441,7 +441,7 @@ int oldnew2c_diff (int context, char *ignore, char *old_fn, char *new_fn, char *
 
   TRACE(fprintf(STDERR,"oldnew2c_diff(%i,%s,%s,%s,%s)\n",context,ignore,old_fn,new_fn,c_diff_fn));
 
-  sprintf(w_diff_fn,"%s%c.difflib-%ld-oldnew2c_diff-w_diff",tmpdir(),DIR_SEP,getpid());
+  sprintf(w_diff_fn,"%s%c.difflib-%ld-oldnew2c_diff-w_diff",tmpdir(),DIR_SEP,(long)getpid());
 
   if(!oldnew2w_diff(context,ignore,old_fn,new_fn,w_diff_fn))
     { UNLINK(w_diff_fn); ERRHNDL(0,"oldnew2w_diff returns 0 in oldnew2c_diff","",1); }
@@ -512,7 +512,7 @@ int lwc_diff2html (char *old_fn, char *new_fn,
       return 0;      
     }
 
-  sprintf(fn_clmn,"%s%c.difflib-%ld-lwc_diff2html-clmn-0-0",tmpdir(),DIR_SEP,getpid());
+  sprintf(fn_clmn,"%s%c.difflib-%ld-lwc_diff2html-clmn-0-0",tmpdir(),DIR_SEP,(long)getpid());
   for(i=0;i<5;i++) 
     { clmn_fn[i]=strdup(fn_clmn); clmn_fn[i][strlen(clmn_fn[i])-3]+=i; }
 
@@ -726,7 +726,7 @@ int u_diff2html (int LWC, char *u_diff_fn, char *html_fn,
 
   TRACE(fprintf(STDERR,"u_diff2html(%i,%s,%s,%s,%s)\n",LWC,u_diff_fn,html_fn,caption,revision));
 
-  sprintf(lwc_diff_fn,"%s%c.difflib-%ld-u_diff2html-lwc_diff",tmpdir(),DIR_SEP,getpid());
+  sprintf(lwc_diff_fn,"%s%c.difflib-%ld-u_diff2html-lwc_diff",tmpdir(),DIR_SEP,(long)getpid());
 
   if(!u_diff2lwc_diff(LWC,u_diff_fn,lwc_diff_fn))
     /* { UNLINK(lwc_diff_fn); ERRHNDL(0,"u_diff2lwc_diff returns 0 in u_diff2html","",1); } */
@@ -747,7 +747,7 @@ int oldnew2html (int LWC, int context, char *ignore, char *old_fn, char *new_fn,
 
   TRACE(fprintf(STDERR,"oldnew2html(%i,%i,%s,%s,%s,%s,%s,%s)\n",LWC,context,ignore,old_fn,new_fn,html_fn,caption,revision));
 
-  sprintf(lwc_diff_fn,"%s%c.difflib-%ld-oldnew2html-lwc_diff",tmpdir(),DIR_SEP,getpid());
+  sprintf(lwc_diff_fn,"%s%c.difflib-%ld-oldnew2html-lwc_diff",tmpdir(),DIR_SEP,(long)getpid());
 
   if(!oldnew2lwc_diff(LWC,context,ignore,old_fn,new_fn,lwc_diff_fn))
     /* { UNLINK(lwc_diff_fn); ERRHNDL(0,"oldnew2lwc_diff returns 0 in oldnew2html","",1); } */
